third_party/blink/web_tests/fast/events/event-should-be-dispatched-for-correct-frame-expected.txt

PASS Event is correctly dispatched
PASS successfullyParsed is true

TEST COMPLETE
